move content of region_helpers.py in region.py
