+++ iris/api/region.py
@@ -14,6 +14,7 @@ try:
 except ImportError:
     from PIL import Image
 
+pyautogui.FAILSAFE = False
 logger = logging.getLogger(__name__)
 
 
@@ -158,6 +159,91 @@ def get_region(region=None, for_ocr=False):
     if is_uhd and not for_ocr:
         grabbed_area = grabbed_area.resize([region.getW(), region.getH()])
     return grabbed_area
+
+
+def generate_region_by_markers(top_left_marker_img=None, bottom_right_marker_img=None):
+    try:
+        wait(top_left_marker_img, 10)
+        exists(bottom_right_marker_img, 10)
+    except Exception as err:
+        logger.error('Unable to find page markers')
+        raise err
+
+    top_left_pos = find(top_left_marker_img)
+    hover(top_left_pos, 0)
+    bottom_right_pos = find(bottom_right_marker_img)
+    hover(bottom_right_pos, 0)
+
+    marker_width, marker_height = get_image_size(bottom_right_marker_img)
+
+    return Region(top_left_pos.x,
+                  top_left_pos.y,
+                  (bottom_right_pos.x + marker_width),
+                  bottom_right_pos.y - top_left_pos.y + marker_height)
+
+
+def create_region_from_patterns(top=None, bottom=None, left=None, right=None, padding_top=None, padding_bottom=None,
+                                padding_left=None, padding_right=None):
+    """
+    Returns a region created from combined area of one or more patterns.
+    Argument names are just for convenience and don't influence outcome.
+    """
+    patterns = []
+    if top:
+        patterns.append(top)
+    if bottom:
+        patterns.append(bottom)
+    if left:
+        patterns.append(left)
+    if right:
+        patterns.append(right)
+
+    if len(patterns) == 0:
+        raise ValueError('One or more patterns required.')
+
+    logger.debug('Creating region from %s pattern(s).' % len(patterns))
+
+    a, b = pyautogui.size()
+    p1 = Location(a, b)
+    p2 = Location(0, 0)
+
+    for pattern in patterns:
+        if exists(pattern, 5):
+            current_pattern = find(pattern)
+            if current_pattern.x < p1.x:
+                p1.x = current_pattern.x
+            if current_pattern.y < p1.y:
+                p1.y = current_pattern.y
+
+            w, h = get_image_size(pattern)
+
+            if current_pattern.x + w > p2.x:
+                p2.x = current_pattern.x + w
+            if current_pattern.y + h > p2.y:
+                p2.y = current_pattern.y + h
+        else:
+            raise FindError('Pattern not found: %s ' % pattern)
+
+    found_region = Region(p1.x, p1.y, p2.x - p1.x, p2.y - p1.y)
+
+    if padding_top or padding_bottom or padding_left or padding_right:
+        logger.debug('Adding padding to region.')
+
+    if padding_top:
+        found_region.y -= padding_top
+        found_region.h += padding_top
+
+    if padding_bottom:
+        found_region.h += padding_bottom
+
+    if padding_left:
+        found_region.x -= padding_left
+        found_region.w += padding_left
+
+    if padding_right:
+        found_region.w += padding_right
+
+    return found_region
 
 
 def _click_at(location=None, clicks=None, duration=None, button=None):
